feat: feedback render type checked
